MM_33223 : Refactored import statements
