+++ app/database/admin/database_manager/index.ts
@@ -1,6 +1,12 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
+import {Database, Q} from '@nozbe/watermelondb';
+import SQLiteAdapter from '@nozbe/watermelondb/adapters/sqlite';
+import {DeviceEventEmitter, Platform} from 'react-native';
+import {FileSystem} from 'react-native-unimodules';
+
+import {MIGRATION_EVENTS, MM_TABLES} from '@constants/database';
 import type {
     ActiveServerDatabase,
     DatabaseConnection,
@@ -9,7 +15,14 @@ import type {
     MigrationEvents,
     Models,
 } from '@typings/database/database';
+import {DatabaseType} from '@typings/database/enums';
+import IServers from '@typings/database/servers';
+import {deleteIOSDatabase, getIOSAppGroupDetails} from '@utils/mattermost_managed';
+
+import DefaultMigration from '../../default/migration';
 import {App, Global, Servers} from '../../default/models';
+import {defaultSchema} from '../../default/schema';
+import ServerMigration from '../../server/migration';
 import {
     Channel,
     ChannelInfo,
@@ -40,26 +53,9 @@ import {
     TermsOfService,
     User,
 } from '../../server/models';
-import {Database, Q} from '@nozbe/watermelondb';
-import {DeviceEventEmitter, Platform} from 'react-native';
-import {MIGRATION_EVENTS, MM_TABLES} from '@constants/database';
-import {deleteIOSDatabase, getIOSAppGroupDetails} from '@utils/mattermost_managed';
-
-import DefaultMigration from '../../default/migration';
-import {FileSystem} from 'react-native-unimodules';
-import IServers from '@typings/database/servers';
-import SQLiteAdapter from '@nozbe/watermelondb/adapters/sqlite';
-import ServerMigration from '../../server/migration';
-import {defaultSchema} from '../../default/schema';
 import {serverSchema} from '../../server/schema';
 
 const {SERVERS} = MM_TABLES.DEFAULT;
-
-// The only two types of databases in the app
-export enum DatabaseType {
-    DEFAULT,
-    SERVER,
-}
 
 class DatabaseManager {
     private activeDatabase: DatabaseInstance;
